added a regridding test file with initial tests for convolution to a specified resolution
